+++ hdk/cl/examples/cl_sde/software/src/sde_lib/sde_mgmt.h
@@ -25,18 +25,20 @@
 #include <stddef.h>
 #include <stdbool.h>
 
-int sde_mgmt_init(int slot_id, enum SDE_EXAMPLE_DIR direction, size_t packet_size, enum SDE_BUFFER_LAYOUT layout);
-int sde_mgmt_init_and_cfg(int slot_id, enum SDE_EXAMPLE_DIR direction, size_t packet_size);
-int sde_mgmt_close(int slot_id);
+#define SDE_PUBLIC __attribute__((visibility("default")))
 
-int sde_mgmt_reset(int slot_id);
-int sde_mgmt_check_status(int slot_id, enum SDE_SUBSYSTEM subsystem);
+SDE_PUBLIC int sde_mgmt_init(int slot_id, enum SDE_EXAMPLE_DIR direction, size_t packet_size, enum SDE_BUFFER_LAYOUT layout);
+SDE_PUBLIC int sde_mgmt_init_and_cfg(int slot_id, enum SDE_EXAMPLE_DIR direction, size_t packet_size);
+SDE_PUBLIC int sde_mgmt_close(int slot_id);
 
-int sde_mgmt_set_dma_buffers(int slot_id, enum SDE_SUBSYSTEM subsystem, struct sde_buffer* sde_buffers, size_t num_buffers);
-int sde_mgmt_cfg(int slot_id);
-int sde_mgmt_wait_desc_credit(int slot_id, enum SDE_SUBSYSTEM subsystem, size_t num_desc);
+SDE_PUBLIC int sde_mgmt_reset(int slot_id);
+SDE_PUBLIC int sde_mgmt_check_status(int slot_id, enum SDE_SUBSYSTEM subsystem);
 
-int sde_mgmt_post_desc(int slot_id, enum SDE_SUBSYSTEM subsystem, size_t* num_desc);
+SDE_PUBLIC int sde_mgmt_set_dma_buffers(int slot_id, enum SDE_SUBSYSTEM subsystem, struct sde_buffer* sde_buffers, size_t num_buffers);
+SDE_PUBLIC int sde_mgmt_cfg(int slot_id);
+SDE_PUBLIC int sde_mgmt_wait_desc_credit(int slot_id, enum SDE_SUBSYSTEM subsystem, size_t num_desc);
+
+SDE_PUBLIC int sde_mgmt_post_desc(int slot_id, enum SDE_SUBSYSTEM subsystem, size_t* num_desc);
 
 struct sde_md {
   uint32_t length;
@@ -45,9 +47,9 @@ struct sde_md {
   uint32_t user_bits[2];
 };
 
-int sde_mgmt_start_read(int slot_id, size_t size);
-int sde_mgmt_read_md(int slot_id, struct sde_md* md);
-int sde_mgmt_read_data(int slot_id, void *data, size_t size);
+SDE_PUBLIC int sde_mgmt_start_read(int slot_id, size_t size);
+SDE_PUBLIC int sde_mgmt_read_md(int slot_id, struct sde_md* md);
+SDE_PUBLIC int sde_mgmt_read_data(int slot_id, void *data, size_t size);
 
-int sde_mgmt_prepare_write(int slot_id, void *data, size_t size);
-int sde_mgmt_write(int slot_id, size_t size);
+SDE_PUBLIC int sde_mgmt_prepare_write(int slot_id, void *data, size_t size);
+SDE_PUBLIC int sde_mgmt_write(int slot_id, size_t size);