+++ sdk_setup.sh
@@ -16,6 +16,9 @@
 # Script must be sourced from a bash shell or it will not work
 # When being sourced $0 will be the interactive shell and $BASH_SOURCE_ will contain the script being sourced
 # When being run $0 and $_ will be the same.
+
+set -x
+
 script=${BASH_SOURCE[0]}
 if [ $script == $0 ]; then
   echo "ERROR: You must source this script"
@@ -38,10 +41,6 @@ fi
 
 sudo rm -f /tmp/sdk_root_env.exp
 typeset -f allow_non_root > /tmp/sdk_root_env.exp
-echo "export AWS_FPGA_SDK_GROUP=${AWS_FPGA_SDK_GROUP}" >> /tmp/sdk_root_env.exp
-echo "export AWS_FPGA_SDK_OTHERS=${AWS_FPGA_SDK_OTHERS}" >> /tmp/sdk_root_env.exp
-echo "export SDK_NON_ROOT_USER=${SDK_NON_ROOT_USER}" >> /tmp/sdk_root_env.exp
-echo "export AWS_FPGA_SDK_OVERRIDE_GROUP=${AWS_FPGA_SDK_OVERRIDE_GROUP}" >> /tmp/sdk_root_env.exp
 sudo chown root:root /tmp/sdk_root_env.exp
 sudo chmod 700 /tmp/sdk_root_env.exp
 
@@ -56,7 +55,6 @@ if ! bash $SDK_DIR/sdk_install.sh; then
 fi
 
 cd sdk/userspace/cython_bindings
-sudo apt install -y python3-venv python3-pip
 python3 -m venv venv
 source venv/bin/activate
 pip install setuptools Cython
@@ -66,4 +64,4 @@ echo "Cython bindings setup complete!"
 
 cd $current_dir
 info_msg "$script_name PASSED"
-
+set +x