+++ shared/bin/set_common_env_vars.sh
@@ -16,10 +16,10 @@
 info_msg "Setting up environment variables"
 
 # Make sure that AWS_FPGA_REPO_DIR is set to the location of this script.
-if [[ ":$AWS_FPGA_REPO_DIR" == ':' ]]; then
+if [[ -z "${AWS_FPGA_REPO_DIR:-}" ]]; then
   debug_msg "AWS_FPGA_REPO_DIR not set so setting to $script_dir"
   export AWS_FPGA_REPO_DIR=$script_dir
-elif [[ $AWS_FPGA_REPO_DIR != $script_dir ]]; then
+elif [[ "$AWS_FPGA_REPO_DIR" != "$script_dir" ]]; then
   info_msg "Changing AWS_FPGA_REPO_DIR from $AWS_FPGA_REPO_DIR to $script_dir"
   export AWS_FPGA_REPO_DIR=$script_dir
 else
@@ -51,11 +51,9 @@ export SDK_DIR=$AWS_FPGA_REPO_DIR/sdk
 #Vitis
 export VITIS_DIR=$AWS_FPGA_REPO_DIR/vitis
 
-# PYTHONPATH
 # Update PYTHONPATH with libraries used for unit testing
 python_lib=$AWS_FPGA_REPO_DIR/shared/lib
-PYTHONPATH=$python_lib:$PYTHONPATH
-export PYTHONPATH
+export PYTHONPATH=$python_lib:${PYTHONPATH:+"$PYTHONPATH"}
 
 # PATH Changes