+++ docs-rtd/source/RELEASE-NOTES.rst
@@ -1,14 +1,24 @@
 F2 Developer Kit Release Notes
 ==============================
 
+.. _v220:
+
+v2.2.0
+------
+
+- Release of Vivado/Vitis 2025.1 Tools on `FPGA Developer AMI 1.18.0 (Ubuntu) <http://aws.amazon.com/marketplace/pp/prodview-yqgciytgjosfw>`__
+- Introduced `MSI-X PCIe Interrupts Guided Example <./sdk/apps/msix-interrupts/README.html>`__
+- Added `Loopback performance test for CL_SDE <./hdk/cl/examples/cl-sde/software/src/README.html>`__
+- `ReadTheDocs navigation improvements <https://awsdocs-fpga-f2.readthedocs-hosted.com/latest/>`__
+
 .. _v212:
 
 v2.1.2
 ------
 
 - Introduced Python Bindings to the SDK
-- Added `documentation <./sdk/userspace/cython-bindings/README.html>` for Python binding usage and setup
-- `Examples <https://github.com/aws/aws-fpga/blob/f2/sdk/userspace/cython_bindings>` demonstrating Python-based FPGA control
+- Added `documentation <./sdk/userspace/cython-bindings/README.html>`__ for Python binding usage and setup
+- `Examples <https://github.com/aws/aws-fpga/blob/f2/sdk/userspace/cython_bindings>`__ demonstrating Python-based FPGA control
 - Added link to instructions for DCV licensing setup. Credit to @morgnza for this update!
 - Added verbiage to DCV setup guide to show where to set virtual display resolution
 - Fix to Bandwidth Calculation
@@ -20,7 +30,7 @@ v2.1.1
 
 - Added global register offset for the SDE IP. See `CL_SDE software examples <./hdk/cl/examples/cl-sde/software/src/README.html>`__.
 - Added `CL_SDE software exmaple <https://github.com/aws/aws-fpga/blob/f2/hdk/cl/examples/cl_sde/software/src/sde_c2h_user_buffers.c>`__ for a user allocated DMA buffer.
-- `Documentation <./hdk/docs/List-AFI-on-Marketplace.html>`__` to assist F2 customers with releasing AFIs and AMIs on the AWS Marketplace.
+- `Documentation <./hdk/docs/List-AFI-on-Marketplace.html>`__ to assist F2 customers with releasing AFIs and AMIs on the AWS Marketplace.
 - `Documentation <./developer-resources/Amazon-DCV-Setup-Guide.html>`__ to assist in creating a virtual desktop based on the FPGA Developer AMI running graphics-intensive applications remotely on Amazon EC2 instances.
 - Fixed the BW calculation and tolerance calculation in the test_hbm_perf_random in the `cl_mem_perf <./hdk/cl/examples/cl-mem-perf/verif/README.html#test-hbm-perf-randomsv-mem>`__.